refactor: drop library API surface, ship as CLI-only
Isotopes is distributed as a CLI binary (`bin: isotopes -> dist/cli.js`).
There are no consumers of `import { ... } from "isotopes"`. Maintaining
the library entry point forced every dead-code review to ask "is this
exported symbol actually unused, or is some hypothetical npm consumer
relying on it?" — a question with no answer.

Changes:
- Delete `src/index.ts` (337 lines of re-exports). Move the lone
  internally-used symbol (`VERSION`) to `src/version.ts`.
- Update `cli.ts` and `api/routes.ts` to import VERSION from the new
  location.
- `package.json`: drop `main`, `types`, `exports` fields. Keep `bin`.
- README: replace the ~100-line "Programmatic API" section with a short
  note that the package is CLI-only.

Net: -442 / +7 lines. From now on, "exported but ungrep'd" = dead code.
