+++ package.json
@@ -3,16 +3,8 @@
   "version": "0.1.0",
   "description": "Lightweight, self-hostable AI agent framework",
   "type": "module",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
   "bin": {
     "isotopes": "dist/cli.js"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
   },
   "scripts": {
     "build": "tsc",
